requirements.txt
praw
azure-storage-blob
openai
pydantic
python-dotenv
PyYAML